Add files via upload
Java-CRUD is a simple database application for universities that allows users to Create, Read, Update, and Delete student information. 

I built this to solidify concepts in my mind of how JSP pages could work with SQL databases and how I could implement a Model-View-Controller architecture to an application. The CSS & JS are kept to a minimum (excuse the bare design here) and the focus was more about how different aspects of the MVC can communicate and interact with one another.